Find median of two sorted arrays
For now, only arrays which combined length is odd, in a specific sequence.
If one array has only one element, which is greater than the highes number
in the other, that number evaluates to be the median :(
Also, if there's only one element, lower than everything else, an
exception happens.